StreamInfo FE refactor
